+++ react/components/Slider/utils/ItemsPerPage.js
@@ -5,12 +5,16 @@ function getItemWidth(slick, maxWidth) {
     ['innerSlider', 'list', 'childNodes', '0', 'childNodes'],
     slick
   )
+
   let itemWidth = null
+
   if (slidesNodeList) {
     const slidesArray = Array.prototype.slice.call(slidesNodeList)
+
     // eslint-disable-next-line array-callback-return
     slidesArray.map(slide => {
       const attributes = Array.prototype.slice.call(slide.attributes)
+
       // eslint-disable-next-line array-callback-return
       attributes.map(attr => {
         // eslint-disable-next-line vtex/prefer-early-return
@@ -21,6 +25,7 @@ function getItemWidth(slick, maxWidth) {
       })
     })
   }
+
   return itemWidth || maxWidth
 }
 
@@ -37,9 +42,11 @@ export default function getItemsPerPage(
   if (slideWidth) {
     const shelfItemWidth = getItemWidth(slick, defaultItemWidth)
     const maxItemsPerPage = Math.floor(slideWidth / shelfItemWidth)
+
     if (actualItemsPerPage >= maxItemsPerPage) {
       return maxItemsPerPage || 1
     }
   }
+
   return actualItemsPerPage
 }