+++ react/typings/vtex.responsive-values.d.ts
@@ -3,6 +3,23 @@ declare module 'vtex.responsive-values' {
   const useResponsiveValues: (input: any) => any
   const useResponsiveValue: <T>(input: MaybeResponsiveInput<T>) => T
 
+  // eslint-disable-next-line no-restricted-syntax
+  enum InputDevices {
+    Mobile = 'mobile',
+    Phone = 'phone',
+    Tablet = 'tablet',
+    Desktop = 'desktop',
+  }
+
+  // eslint-disable-next-line no-restricted-syntax
+  enum InputDevices {
+    MOBILE = 'mobile',
+    PHONE = 'phone',
+    TABLET = 'tablet',
+    DESKTOP = 'desktop',
+  }
+
+  // eslint-disable-next-line no-restricted-syntax
   enum InputDevices {
     mobile = 'mobile',
     phone = 'phone',
@@ -10,6 +27,7 @@ declare module 'vtex.responsive-values' {
     desktop = 'desktop',
   }
 
+  // eslint-disable-next-line no-restricted-syntax
   enum OutputDevices {
     phone = 'phone',
     tablet = 'tablet',